bugs fixed

+++ src/main/java/pl/dyplom/dyplom/domain/ProductQuantity.java
@@ -13,7 +13,7 @@ public class ProductQuantity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private Long qty;
+    private int qty;
 
     @OneToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "ordered_product_id")
